Tweak death effect more

+++ src/audio_visual/raycast.c
@@ -37,7 +37,7 @@ vec handle_ray(const DataRaycast d) {
 	const double smallest_wall_y = (double) wall_dest.y - (wall_h * (point_height - 1)); // = wall_top
 
 	if (d.first_wall_hit) update_val_buffers(d.screen_x, smallest_wall_y,
-		wall_dest.y + wall_dest.h, corrected_dist, cos_beta, d.dir);
+		(double) (wall_dest.y + wall_dest.h), corrected_dist, cos_beta, d.dir);
 
 	const Sprite wall_sprite = current_level.walls[d.point - 1];
 	const SDL_Rect mipmap_crop = get_mipmap_crop_from_dist(wall_sprite.size, corrected_dist);
